Upgrade to Python 3.8 (#1103)
Python 2 is EOL as of Jan 1, 2019 so we should not support it anymore.
Everything that uses Python (collectd-python, the SignalFx python
subprocess runner) will use Python 3, and it will be the Python runtime
bundled with the agent.  The integration test suite was already using
Python 3.

Also upgraded all collectd python monitors to support Python 3 and
pointed to the new tags in collectd-plugins.yaml.

Removed `collectd/haproxy` and `collectd/docker` monitors since they are
deprecated and have equivalent Go-based monitors.  I left
`collectd/elasticsearch` for now since it is used heavily internally,
but it can probably be removed in the next major release.